refactor(post): 优化帖子服务代码结构及功能实现
- 统一点赞、取消点赞、收藏、取消收藏接口返回结果格式
- 使用枚举封装操作结果状态和响应，简化代码
- 更改帖子实体主键类型及字段命名，新增逻辑删除字段
- 实现批量获取帖子详情接口，支持缓存优先查询数据库补全
- 重构帖子详情查询，集成推荐服务和缓存访问逻辑
- 删除多余共享功能及相关接口代码
- 实现异步处理帖子行为消息消费，分离同步更新逻辑
- 完善服务间Feign调用，修正社团ID类型，新增推荐服务Feign
- 统一缓存的帖子字段名及过期时间配置
- 优化Redis缓存的点赞与收藏状态判断及缓存加载逻辑
- 调整日志及异常处理，增强可维护性和代码整洁度
- 文件上传相关控制器注解合并优化，移除无用代码和依赖

+++ services/post-service/src/main/java/org/example/feign/ClubFeign.java
@@ -10,7 +10,6 @@ import java.util.Set;
 @FeignClient(name = "club-service")
 public interface ClubFeign {
 
-
     @PostMapping("/api/club/batch")
-    Map<Integer, ClubBriefDto> getClubInfosByIds(Set<Integer> clubIds);
+    Map<Long, ClubBriefDto> getClubInfosByIds(Set<Long> clubIds);
 }
